refactor(native-spans): renamed the "remote spans" module as "native spans"
