+++ packages/plugin-react-native-native-spans/ios/BugsnagNativeSpans.h
@@ -1,0 +1,19 @@
+#import <React/RCTBridgeModule.h>
+
+#ifdef RCT_NEW_ARCH_ENABLED
+#import <BugsnagNativeSpansSpec/BugsnagNativeSpansSpec.h>
+#endif
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface BugsnagNativeSpans : NSObject <RCTBridgeModule>
+
+@end
+
+#ifdef RCT_NEW_ARCH_ENABLED
+@interface BugsnagNativeSpans () <NativeBugsnagNativeSpansSpec>
+
+@end
+#endif
+
+NS_ASSUME_NONNULL_END
