+++ packages/plugin-react-native-native-spans/ios/BugsnagNativeSpansPlugin+Private.h
@@ -1,0 +1,10 @@
+#import "BugsnagNativeSpansPlugin.h"
+
+@interface BugsnagNativeSpansPlugin ()
+
++ (id)singleton;
+
+@property(nonatomic, readonly) NSMutableDictionary *spansByName;
+@property(nonatomic, readonly) NSMutableDictionary *spansById;
+
+@end
