+++ packages/plugin-react-native-native-spans/ios/BugsnagNativeSpansPlugin.h
@@ -1,0 +1,10 @@
+#import <Foundation/Foundation.h>
+#import <BugsnagPerformance/BugsnagPerformancePlugin.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface BugsnagNativeSpansPlugin: NSObject <BugsnagPerformancePlugin>
+
+@end
+
+NS_ASSUME_NONNULL_END
