+++ packages/plugin-react-native-native-spans/tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "extends": "../../tsconfig.json",
+    "include": ["lib/*"],
+    "exclude": [
+      "lib/NativeBugsnagNativeSpans.ts"
+    ],
+    "compilerOptions": {
+      "target": "ES2020"
+    }
+  }
