--- packages/plugin-react-native-remote-spans/android/src/oldarch/java/com/bugsnag/reactnative/performance/remotespans/BugsnagRemoteSpansPackage.java
@@ -1,24 +1,0 @@
-package com.bugsnag.reactnative.performance.remotespans;
-
-import com.facebook.react.ReactPackage;
-import com.facebook.react.bridge.NativeModule;
-import com.facebook.react.bridge.ReactApplicationContext;
-import com.facebook.react.uimanager.ViewManager;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-
-public class BugsnagRemoteSpansPackage implements ReactPackage {
-  @Override
-  public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
-    List<NativeModule> modules = new ArrayList<>();
-    modules.add(new BugsnagRemoteSpansModule(reactContext));
-    return modules;
-  }
-
-  @Override
-  public List<ViewManager> createViewManagers(ReactApplicationContext reactContext) {
-    return Collections.emptyList();
-  }
-}
