--- packages/plugin-react-native-remote-spans/ios/BugsnagRemoteSpansPlugin+Private.h
@@ -1,10 +1,0 @@
-#import "BugsnagRemoteSpansPlugin.h"
-
-@interface BugsnagRemoteSpansPlugin ()
-
-+ (id)singleton;
-
-@property(nonatomic, readonly) NSMutableDictionary *spansByName;
-@property(nonatomic, readonly) NSMutableDictionary *spansById;
-
-@end
