--- packages/plugin-react-native-remote-spans/ios/BugsnagRemoteSpansPlugin.h
@@ -1,10 +1,0 @@
-#import <Foundation/Foundation.h>
-#import <BugsnagPerformance/BugsnagPerformancePlugin.h>
-
-NS_ASSUME_NONNULL_BEGIN
-
-@interface BugsnagRemoteSpansPlugin: NSObject <BugsnagPerformancePlugin>
-
-@end
-
-NS_ASSUME_NONNULL_END
