--- packages/plugin-react-native-remote-spans/tsconfig.json
@@ -1,10 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "include": ["lib/*"],
-    "exclude": [
-      "lib/NativeBugsnagRemoteSpans.ts"
-    ],
-    "compilerOptions": {
-      "target": "ES2020"
-    }
-  }
