+++ test/react-native/features/fixtures/scenario-launcher/ios/ScenarioLauncher.mm
@@ -5,7 +5,7 @@
 #ifdef NATIVE_INTEGRATION
 #import <BugsnagPerformance/BugsnagPerformance.h>
 #import <BugsnagPerformance/BugsnagPerformanceConfiguration+Private.h>
-#import "BugsnagRemoteSpansPlugin.h"
+#import "BugsnagNativeSpansPlugin.h"
 #endif
 
 @implementation ScenarioLauncher {
@@ -144,7 +144,7 @@ RCT_EXPORT_METHOD(startNativePerformance:(NSDictionary *)configuration resolve:(
     config.internal.autoTriggerExportOnBatchSize = 1;
     config.internal.clearPersistenceOnStart = YES;
 
-    [config addPlugin:[BugsnagRemoteSpansPlugin new]];
+    [config addPlugin:[BugsnagNativeSpansPlugin new]];
 
     [BugsnagPerformance startWithConfiguration:config];
     resolve(nil);
